Chunk data operations made fast.
Added a function for the data size of a tag that can measure the size much faster for a tag than by calling get_data().size(). In addition, added a new get_data function which takes a byte_stream by reference, and used it in chunk writing, thereby reducing the number of creations and destructions of vectors. Changed the write_stream function of the byte_stream to directly read bytes from memory in the endiandness needed, without creating helper vectors. Improvements in performance are up to 50-fold for writing and up to 100-fold for chunk generation.

(cherry picked from commit 18f897ca67e959283b64ca819f4f230749a156b0)